% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{bagging}
\alias{bagging}
\title{Bagging for Linear, Logistic, Ridge, Lasso and Elastic net Models}
\usage{
bagging(X, y, model_type = "linear", R = 100, alpha = NULL, lambda = NULL)
}
\arguments{
\item{X}{Predictor matrix.}

\item{y}{Response variable.}

\item{model_type}{Character string indicating the type of model to use ("linear", "logistic", "ridge", "lasso", "elastic_net").}

\item{R}{Integer specifying the number of bootstrap samples to draw.}

\item{alpha}{Ridge/lasso mixing parameter (0 for ridge, 1 for lasso) for elastic net.}

\item{lambda}{Regularization parameter for ridge and lasso regression.}
}
\value{
A list containing the final predicted values and variable importance scores.
}
\description{
This function performs bagging (bootstrap aggregation) for various regression and classification models, returning final predicted values and variable importance scores where applicable.
}
\examples{
n <- 100
p <- 20
X <- as.matrix(matrix(rnorm(n * p), nrow = n))
beta <- c(1, rep(0, p - 1))
y <- X \%*\% beta + rnorm(n)
predictions <- bagging(X,y,model_type = "lasso",R =100,lambda=0.2)
print(predictions)

}
