% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{elastic_net}
\alias{elastic_net}
\title{Elastic Net}
\usage{
elastic_net(
  X,
  y,
  bagging = FALSE,
  prescreen = FALSE,
  alpha = 0.5,
  lambda = 0.1
)
}
\arguments{
\item{X}{Predictor matrix.}

\item{y}{Response variable.}

\item{bagging}{Logical indicating whether to perform bagging.}

\item{prescreen}{Logical indicating whether to pre-screen predictors.}

\item{alpha}{Ridge/lasso mixing parameter (0 for ridge, 1 for lasso).}

\item{lambda}{Regularization parameter.}
}
\value{
The fitted elastic net model.
}
\description{
The elastic_net function implements elastic net regression using the glmnet package. Elastic net is a regularized linear regression technique that combines the properties of both ridge regression and lasso regression. It can perform feature selection and coefficient shrinkage simultaneously.
}
\examples{
#' # Generate sample data for a elastic net  problem
n <- 100 # Number of observations
p <- 10 # Number of predictors
X <- as.matrix(matrix(rnorm(n * p), nrow = n)) # Generate predictor matrix
beta <- c(1, rep(0, p - 1)) # True coefficients (first one is 1, others are 0)
y <- X \%*\% beta + rnorm(n) # Generate the response variable

## Linear regression without bagging or prescreening
# Fit the Linear regression model without bagging or prescreening
model1 <- elastic_net(X, y, bagging = FALSE, prescreen = FALSE, lambda = 0.1)

# Make predictions on the original data
predictions1 <- predict(model1, newx = X)

print(head(predictions1))

}
