% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear Regression}
\usage{
linear_regression(
  X,
  y,
  bagging = FALSE,
  prescreen = FALSE,
  alpha = 0,
  lambda = 0.1
)
}
\arguments{
\item{X}{Predictor matrix.}

\item{y}{Response variable.}

\item{bagging}{Logical indicating whether to perform bagging.}

\item{prescreen}{Logical indicating whether to pre-screen predictors.}

\item{alpha}{Ridge/lasso mixing parameter (0 for ridge, 1 for lasso).}

\item{lambda}{Regularization parameter.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
The fitted linear regression model.
}
\description{
The linear_regression function implements linear regression using the glmnet package. Linear regression is a widely used statistical model for predicting a continuous response variable based on one or more predictor variables.
}
\examples{
# Generate sample data for a linear regression problem
n <- 100 # Number of observations
p <- 10 # Number of predictors
X <- as.matrix(matrix(rnorm(n * p), nrow = n)) # Generate predictor matrix
beta <- c(1, rep(0, p - 1)) # True coefficients (first one is 1, others are 0)
y <- X \%*\% beta + rnorm(n) # Generate the response variable

## Linear regression without bagging or prescreening
# Fit the Linear regression model without bagging or prescreening
model1 <- linear_regression(X, y, bagging = FALSE, prescreen = FALSE, alpha = 0, lambda = 0.1)

# Make predictions on the original data
predictions1 <- predict(model1, newx = X)

print(head(predictions1)



}
