% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescreen.R
\name{prescreen_predictors}
\alias{prescreen_predictors}
\title{Prescreen Top K Predictors}
\usage{
prescreen_predictors(X, y, K = 7)
}
\arguments{
\item{X}{A matrix where each column represents a predictor variable.}

\item{y}{A matrix with a single column representing the response variable.}

\item{K}{An integer specifying the number of top predictors to select (default is 7).}
}
\value{
A matrix containing the top K predictors selected based on the Random Forest
  variable importance scores.
}
\description{
This function prescreens the top K most informative predictors based on
the variable importance scores from a Random Forest model.
}
\examples{
# Generate some example data
set.seed(123)
n <- 100
p <- 2000
X <- matrix(rnorm(n * p), nrow = n)
y <- matrix(rnorm(n), ncol = 1)
# Prescreen top 5 predictors
top_predictors <- prescreen_predictors(X, as.vector(y), K = 20)

print(head(top_predictors))
#For binary response
y_bin <- matrix(sample(c(0, 1), n, replace = TRUE), ncol = 1)
top_predictors_bin <- prescreen_predictors(X,as.factor(y_bin), K = 20)

print(head(top_predictors_bin))

}
